using react toolkit and organized the code so its more manageable
