+++ package.json
@@ -25,6 +25,7 @@
     "@ant-design/icons": "^4.4.0",
     "@bbp/nexus-sdk": "1.3.15",
     "@bbp/react-nexus": "1.3.15",
+    "@reduxjs/toolkit": "^1.8.5",
     "@sentry/browser": "6.13.1",
     "@types/cytoscape": "^3.19.4",
     "@types/lodash": "^4.14.182",
@@ -76,7 +77,7 @@
     "react-pdf": "5.3.0",
     "react-player": "^2.9.0",
     "react-query": "^3.13.4",
-    "react-redux": "^7.1.1",
+    "react-redux": "^8.0.2",
     "react-router": "^5.1.2",
     "react-router-dom": "^5.1.2",
     "react-spring": "^8.0.19",
@@ -122,7 +123,7 @@
     "@types/react-dom": "^16.9.0",
     "@types/react-helmet": "^5.0.7",
     "@types/react-infinite-scroll-component": "^4.2.5",
-    "@types/react-redux": "^7.0.3",
+    "@types/react-redux": "^7.1.24",
     "@types/react-router": "^5.1.13",
     "@types/react-router-dom": "^5.1.7",
     "@types/showdown": "^1.9.3",
