+++ src/constants/index.ts
@@ -1,0 +1,50 @@
+// Report Constants
+export const REPORT_CATEGORIES = {
+  circuit: [
+    'Anatomical',
+    'Connectivity',
+    'Volumetric',
+    'Morphometric',
+    'Synapse',
+  ],
+  simulation: ['Spiking', 'Soma voltage', 'LFP', 'VSD', 'Plasticity'],
+};
+
+export const REPORT_TYPES = ['Validation', 'Prediction', 'Analysis'];
+
+export const DEFAULT_ANALYSIS_DATA_SPARQL_QUERY = `PREFIX s:<http://schema.org/>
+PREFIX prov:<http://www.w3.org/ns/prov#>
+PREFIX nsg:<https://neuroshapes.org/>
+PREFIX nxv:<https://bluebrain.github.io/nexus/vocabulary/>
+SELECT ?container_resource_id  ?container_resource_type ?container_resource_name ?analysis_report_id ?analysis_report_name ?analysis_report_description ?created_by ?created_at ?self
+WHERE {
+  OPTIONAL {
+    BIND(<{resourceId}> as ?container_resource_id) .
+    BIND(<{resourceId}> as ?self) .
+    ?derivation_id        ^prov:derivation       ?analysis_report_id .
+    ?derivation_id        nsg:entity             ?container_resource_id .
+    OPTIONAL {
+      ?container_resource_id        nsg:name                   ?container_resource_name .
+      ?container_resource_id        nsg:type                   ?container_resource_type .
+    }
+    ?analysis_report_id    nsg:name            ?analysis_report_name .
+    ?analysis_report_id    nsg:description       ?analysis_report_description .
+    ?analysis_report_id nxv:createdBy ?created_by .
+    ?analysis_report_id nxv:createdAt ?created_at .
+  }
+  OPTIONAL {
+    BIND(<{resourceId}> as ?analysis_report_id) .
+    BIND(<{resourceId}> as ?self) .
+    ?derivation_id        ^prov:derivation       ?analysis_report_id .
+    ?derivation_id        nsg:entity             ?container_resource_id .
+    OPTIONAL {
+      ?container_resource_id        nsg:name                   ?container_resource_name .
+      ?container_resource_id        nsg:type                   ?container_resource_type .
+    }
+    ?analysis_report_id    nsg:name            ?analysis_report_name .
+    ?analysis_report_id    nsg:description       ?analysis_report_description .
+    ?analysis_report_id nxv:createdBy ?created_by .
+    ?analysis_report_id nxv:createdAt ?created_at .
+  }
+}
+LIMIT 1000`;
