+++ src/shared/containers/AnalysisPlugin/AnalysisPluginContainer.tsx
@@ -4,10 +4,6 @@ import * as React from 'react';
 import { useMutation, useQuery, useQueryClient } from 'react-query';
 import { useSelector } from 'react-redux';
 import { RootState } from '../../store/reducers';
-import AnalysisPlugin, {
-  AnalysisReport,
-  Asset,
-} from '../../components/AnalysisPlugin/AnalysisPlugin';
 import { sparqlQueryExecutor } from '../../utils/querySparqlView';
 import { Image } from 'antd';
 import FileUploadContainer from '../FileUploadContainer';
@@ -17,43 +13,29 @@ import { useHistory, useLocation } from 'react-router';
 import ImageFileInfo from '../../components/FileInfo/ImageFileInfo';
 import { PDFThumbnail } from '../../../shared/components/Preview/PDFPreview';
 import PDFFileInfo from '../../../shared/components/FileInfo/PDFFileInfo';
+import analysisUIReducer from '../../slices/plugins/report';
+import AnalysisPlugin from '../../../shared/components/AnalysisPlugin/AnalysisPlugin';
+import {
+  setReportResourceType,
+  editReport,
+  addReport,
+  selectAsset,
+  setSelectedReportFirstLoad,
+  changeSelectedReports,
+  changeAnalysisName,
+  changeAnalysisCategories,
+  changeAnalysisTypes,
+  openFileUploadDialog,
+  closeFileUploadDialog,
+} from '../../slices/plugins/report';
 
-export const DEFAULT_ANALYSIS_DATA_SPARQL_QUERY = `PREFIX s:<http://schema.org/>
-PREFIX prov:<http://www.w3.org/ns/prov#>
-PREFIX nsg:<https://neuroshapes.org/>
-PREFIX nxv:<https://bluebrain.github.io/nexus/vocabulary/>
-SELECT ?container_resource_id  ?container_resource_type ?container_resource_name ?analysis_report_id ?analysis_report_name ?analysis_report_description ?created_by ?created_at ?self
-WHERE {
-  OPTIONAL {
-    BIND(<{resourceId}> as ?container_resource_id) .
-    BIND(<{resourceId}> as ?self) .
-    ?derivation_id        ^prov:derivation       ?analysis_report_id .
-    ?derivation_id        nsg:entity             ?container_resource_id .
-    OPTIONAL {
-      ?container_resource_id        nsg:name                   ?container_resource_name .
-      ?container_resource_id        nsg:type                   ?container_resource_type .
-    }
-    ?analysis_report_id    nsg:name            ?analysis_report_name .
-    ?analysis_report_id    nsg:description       ?analysis_report_description .
-    ?analysis_report_id nxv:createdBy ?created_by .
-    ?analysis_report_id nxv:createdAt ?created_at .
-  }
-  OPTIONAL {
-    BIND(<{resourceId}> as ?analysis_report_id) .
-    BIND(<{resourceId}> as ?self) .
-    ?derivation_id        ^prov:derivation       ?analysis_report_id .
-    ?derivation_id        nsg:entity             ?container_resource_id .
-    OPTIONAL {
-      ?container_resource_id        nsg:name                   ?container_resource_name .
-      ?container_resource_id        nsg:type                   ?container_resource_type .
-    }
-    ?analysis_report_id    nsg:name            ?analysis_report_name .
-    ?analysis_report_id    nsg:description       ?analysis_report_description .
-    ?analysis_report_id nxv:createdBy ?created_by .
-    ?analysis_report_id nxv:createdAt ?created_at .
-  }
-}
-LIMIT 1000`;
+import {
+  Asset,
+  AnalysesState,
+  AnalysisReport,
+  AnalysisPluginContainerProps,
+  AnalysisAssetSparqlQueryRowResult,
+} from '../../types/plugins/report';
 
 async function fetchImageObjectUrl(
   nexus: NexusClient,
@@ -75,89 +57,6 @@ async function fetchImageObjectUrl(
   });
   return URL.createObjectURL(blob);
 }
-
-type AnalysisPluginContainerProps = {
-  orgLabel: string;
-  projectLabel: string;
-  resourceId: string;
-};
-
-export enum ActionType {
-  RESCALE = 'rescale',
-  EDIT_ANALYSIS_REPORT = 'edit_analysis_report',
-  INITIALIZE = 'initialize',
-  SELECT_ASSET = 'select_asset',
-  CHANGE_SELECTED_ANALYSIS_REPORTS = 'change_selected_analysis_reports',
-  CHANGE_ANALYSIS_NAME = 'change_analysis_name',
-  SAVE_NEW_ANALYSIS_REPORT = 'save_new_analysis_report',
-  CHANGE_ANALYSIS_DESCRIPTION = 'change_analysis_description',
-  CHANGE_ANALYSIS_CATEGORIES = 'change_analysis_categories',
-  CHANGE_ANALYSIS_TYPES = 'change_analysis_types',
-  OPEN_FILE_UPLOAD_DIALOG = 'open_file_upload_dialog',
-  CLOSE_FILE_UPLOAD_DIALOG = 'close_file_upload_dialog',
-  ADD_ANALYSIS_REPORT = 'add_analysis_report',
-  SET_SELECTED_REPORT_ON_FIRST_LOAD = 'set_selected_report_on_first_load',
-  SET_ANALYSIS_RESOURCE_TYPE = 'set_analysis_resource_type',
-}
-
-export type AnalysesAction =
-  | {
-      type: ActionType.SET_ANALYSIS_RESOURCE_TYPE;
-      payload: {
-        resourceType: 'report_container' | 'individual_report';
-        containerId?: string;
-        containerName?: string;
-      };
-    }
-  | { type: ActionType.RESCALE; payload: number }
-  | {
-      type: ActionType.SET_SELECTED_REPORT_ON_FIRST_LOAD;
-      payload?: { analysisReportId: string };
-    }
-  | {
-      type: ActionType.EDIT_ANALYSIS_REPORT;
-      payload: {
-        analysisId: string;
-        analaysisName: string;
-        analysisDescription?: string;
-        analysisCategories?: string[];
-        analysisTypes?: string[];
-      };
-    }
-  | {
-      type: ActionType.SAVE_NEW_ANALYSIS_REPORT;
-      payload: {
-        name: string;
-        description?: string;
-        categories?: string[];
-        types?: string[];
-      };
-    }
-  | {
-      type: ActionType.INITIALIZE;
-      payload: { scale: number; analysisReportId?: string[] };
-    }
-  | { type: ActionType.SELECT_ASSET; payload: { assetId: string } }
-  | {
-      type: ActionType.CHANGE_SELECTED_ANALYSIS_REPORTS;
-      payload: { analysisReportIds: string[] };
-    }
-  | { type: ActionType.CHANGE_ANALYSIS_NAME; payload: { name: string } }
-  | {
-      type: ActionType.CHANGE_ANALYSIS_DESCRIPTION;
-      payload: { description: string };
-    }
-  | {
-      type: ActionType.CHANGE_ANALYSIS_CATEGORIES;
-      payload: { categories: string[] };
-    }
-  | {
-      type: ActionType.CHANGE_ANALYSIS_TYPES;
-      payload: { types: string[] };
-    }
-  | { type: ActionType.OPEN_FILE_UPLOAD_DIALOG }
-  | { type: ActionType.CLOSE_FILE_UPLOAD_DIALOG }
-  | { type: ActionType.ADD_ANALYSIS_REPORT };
 
 const AnalysisPluginContainer = ({
   orgLabel,
@@ -200,29 +99,6 @@ const AnalysisPluginContainer = ({
     '{resourceId}',
     resourceId
   );
-
-  type AnalysisAssetSparqlQueryRowResult = {
-    id: string;
-    key: string;
-    container_resource_id: string;
-    container_resource_name: string;
-    container_resource_type: string;
-    analysis_report_id: string;
-    analysis_report_name: string;
-    analysis_report_description: string;
-    analysis_report_categories: string[];
-    analysis_report_types: string[];
-    created_by: string;
-    created_at: string;
-    asset_name: string;
-    asset_description: string;
-    asset_content_url: string;
-    asset_encoding_format: string;
-    self: {
-      type: string;
-      value: string;
-    };
-  };
 
   const fetchAnalysisData = async (
     viewSelfId: string,
@@ -314,17 +190,15 @@ const AnalysisPluginContainer = ({
     {
       onSuccess: data => {
         if (!hasInitializedSelectedReports) {
-          dispatch({
-            type: ActionType.SET_SELECTED_REPORT_ON_FIRST_LOAD,
-            payload: {
+          dispatch(
+            setSelectedReportFirstLoad({
               analysisReportId:
                 data.length > 0 && data[0].id !== undefined ? data[0].id : '',
-            },
-          });
+            })
+          );
         }
-        dispatch({
-          type: ActionType.SET_ANALYSIS_RESOURCE_TYPE,
-          payload: {
+        dispatch(
+          setReportResourceType({
             resourceType:
               data.length > 0 && data[0].id === resourceId
                 ? 'individual_report'
@@ -337,8 +211,8 @@ const AnalysisPluginContainer = ({
               data.length > 0 && data[0].containerName !== ''
                 ? data[0].containerName
                 : undefined,
-          },
-        });
+          })
+        );
       },
     }
   );
@@ -489,7 +363,7 @@ const AnalysisPluginContainer = ({
           },
         };
       });
-      
+
       if (data.id) {
         const resource = (await nexus.Resource.get(
           orgLabel,
@@ -556,10 +430,7 @@ const AnalysisPluginContainer = ({
           queryClient.invalidateQueries(['analysis']),
           queryClient.invalidateQueries(['analysesImages']),
         ]).then(() => {
-          dispatch({
-            type: ActionType.CHANGE_SELECTED_ANALYSIS_REPORTS,
-            payload: { analysisReportIds: [resource['@id']] },
-          });
+          dispatch(changeSelectedReports({ analysisReportIds: [resource['@id']] }));
         });
       },
     }
@@ -629,10 +500,7 @@ const AnalysisPluginContainer = ({
           queryClient.invalidateQueries(['analysis']),
           queryClient.invalidateQueries(['analysesImages']),
         ]).then(() => {
-          dispatch({
-            type: ActionType.CHANGE_SELECTED_ANALYSIS_REPORTS,
-            payload: { analysisReportIds: resourceIds ? resourceIds : [] },
-          });
+          dispatch(changeSelectedReports({ analysisReportIds: resourceIds ? resourceIds : [] }));
         });
       },
     }
@@ -686,24 +554,6 @@ const AnalysisPluginContainer = ({
 
   const DEFAULT_SCALE = 50;
 
-  type AnalysesState = {
-    analysisResourceType: 'report_container' | 'individual_report';
-    containerId?: string;
-    containerName?: string;
-    containerType?: string;
-    imagePreviewScale: number;
-    mode: 'view' | 'edit' | 'create';
-    hasInitializedSelectedReports: boolean;
-    selectedAnalysisReports?: string[];
-    currentlyBeingEditedAnalysisReportId?: string;
-    currentlyBeingEditingAnalysisReportName?: string;
-    currentlyBeingEditedAnalysisReportCategories?: string[];
-    currentlyBeingEditedAnalysisReportTypes?: string[];
-    currentlyBeingEditedAnalysisReportDescription?: string;
-    selectedAssets?: string[];
-    isUploadAssetDialogOpen?: boolean;
-  };
-
   const initState = ({
     mode,
     analysisResourceType,
@@ -718,132 +568,6 @@ const AnalysisPluginContainer = ({
       selectedAnalysisReports,
       imagePreviewScale: scale,
     };
-  };
-
-  const analysisUIReducer = (
-    state: AnalysesState,
-    action: AnalysesAction
-  ): AnalysesState => {
-    switch (action.type) {
-      case ActionType.SET_ANALYSIS_RESOURCE_TYPE:
-        return {
-          ...state,
-          analysisResourceType: action.payload.resourceType,
-          containerId: action.payload.containerId,
-          containerName: action.payload.containerName,
-        };
-      case ActionType.RESCALE:
-        return { ...state, imagePreviewScale: action.payload };
-      case ActionType.EDIT_ANALYSIS_REPORT:
-        return {
-          ...state,
-          mode: 'edit',
-          selectedAnalysisReports: [action.payload.analysisId],
-          currentlyBeingEditedAnalysisReportId: action.payload.analysisId,
-          currentlyBeingEditingAnalysisReportName: action.payload.analaysisName,
-          currentlyBeingEditedAnalysisReportDescription:
-            action.payload.analysisDescription,
-        };
-      case ActionType.ADD_ANALYSIS_REPORT:
-        return {
-          ...state,
-          mode: 'create',
-          currentlyBeingEditingAnalysisReportName: '',
-          currentlyBeingEditedAnalysisReportDescription: '',
-          currentlyBeingEditedAnalysisReportCategories: [],
-          currentlyBeingEditedAnalysisReportTypes: [],
-        };
-      case ActionType.SAVE_NEW_ANALYSIS_REPORT:
-        return {
-          ...state,
-          currentlyBeingEditingAnalysisReportName: action.payload.name,
-          currentlyBeingEditedAnalysisReportDescription:
-            action.payload.description,
-          currentlyBeingEditedAnalysisReportCategories:
-            action.payload.categories,
-          currentlyBeingEditedAnalysisReportTypes: action.payload.types,
-        };
-      case ActionType.SELECT_ASSET:
-        state.selectedAssets = state.selectedAssets ? state.selectedAssets : [];
-        const selectedId = state.selectedAssets?.findIndex(
-          a => a === action.payload.assetId
-        );
-
-        if (
-          state.selectedAssets &&
-          selectedId !== undefined &&
-          selectedId > -1
-        ) {
-          const selectedCopy = [...state.selectedAssets];
-          selectedCopy.splice(selectedId, 1);
-          return {
-            ...state,
-            selectedAssets: selectedCopy,
-          };
-        }
-        return {
-          ...state,
-          selectedAssets: [...state.selectedAssets, action.payload.assetId],
-        };
-      case ActionType.SET_SELECTED_REPORT_ON_FIRST_LOAD:
-        return {
-          ...state,
-          selectedAnalysisReports:
-            action.payload && action.payload.analysisReportId !== ''
-              ? [action.payload.analysisReportId]
-              : [],
-          hasInitializedSelectedReports: true,
-        };
-      case ActionType.CHANGE_SELECTED_ANALYSIS_REPORTS:
-        return {
-          ...state,
-          mode: 'view',
-          selectedAnalysisReports: action.payload.analysisReportIds,
-          selectedAssets: [],
-        };
-      case ActionType.INITIALIZE:
-        return initState({
-          mode: 'view',
-          analysisResourceType: 'report_container',
-          hasInitializedSelectedReports: true,
-          selectedAnalysisReports: action.payload.analysisReportId,
-          imagePreviewScale: action.payload.scale,
-        });
-      case ActionType.CHANGE_ANALYSIS_NAME:
-        return {
-          ...state,
-          currentlyBeingEditingAnalysisReportName: action.payload.name,
-        };
-      case ActionType.CHANGE_ANALYSIS_DESCRIPTION:
-        return {
-          ...state,
-          currentlyBeingEditedAnalysisReportDescription:
-            action.payload.description,
-        };
-      case ActionType.CHANGE_ANALYSIS_CATEGORIES:
-        return {
-          ...state,
-          currentlyBeingEditedAnalysisReportCategories:
-            action.payload.categories,
-        };
-      case ActionType.CHANGE_ANALYSIS_TYPES:
-        return {
-          ...state,
-          currentlyBeingEditedAnalysisReportTypes: action.payload.types,
-        };
-      case ActionType.OPEN_FILE_UPLOAD_DIALOG:
-        return {
-          ...state,
-          isUploadAssetDialogOpen: true,
-        };
-      case ActionType.CLOSE_FILE_UPLOAD_DIALOG:
-        return {
-          ...state,
-          isUploadAssetDialogOpen: false,
-        };
-      default:
-        throw new Error();
-    }
   };
 
   const [
@@ -1002,11 +726,9 @@ const AnalysisPluginContainer = ({
             currentlyBeingEditingAnalysisReportName
           }
           selectedAssets={selectedAssets}
+          dispatch={() => dispatch}
           selectedAnalysisReports={selectedAnalysisReports}
           isUploadAssetDialogOpen={isUploadAssetDialogOpen}
-          dispatch={(action: AnalysesAction) => {
-            dispatch(action);
-          }}
           onClickRelatedResource={(resourceId: string) =>
             handleClickAnalysisResource(orgLabel, projectLabel, resourceId)
           }
